refactor(padi): carve connectPadi into @kolu/padi/dial (the shared dial kit)
W2.3 movement 1 — pure motion. `connectPadi` + `dialPadiHello` (state-root →
digest → socket resolve · control-core handshake · typed skew refusal · the
typed padiSurface client) move out of the kolu-server binder into a new
`@kolu/padi/dial` entry, so kolu-server and the coming `padi-tui` share ONE
client-side dial kit — the kaval precedent: a daemon's package owns the dial
kit its clients share.

Supervision stays binder-only in `server/src/padiBinding.ts`: the version
ORDERING (`isBinderNewer`), the drivers, `adoptOrSpawnOrRefuse` convergence,
the reconnect-mirror session, the legacy-kaval hint, and the re-serve —
everything that mutates padi's lifecycle, never a mere dial.

The seal's allowed `@kolu/padi` entries grow by exactly `/dial` (criterion h:
server imports only @kolu/padi/{assembly,surface,dial,log}). The dial's client
types are spelled from padi's own oRPC deps (`ContractRouterClient<C,
ClientRetryPluginContext>`, structurally identical to surface-nix-host's
`AgentClient`) so the kit adds no surface-nix-host edge to padi's closure.

No behavior change: @kolu/padi + kolu-server typecheck, the seal (6/6, /dial
added), padiConverge (14/14), and padiBinding (12/12, real padi spawn/adopt
through the carved connectPadi) all green.
